+++ CHANGELOG.md
@@ -532,7 +532,7 @@ All versions prior to 0.0.9 are untracked.
 
 <!-- Release URLs -->
 [Unreleased]: https://github.com/pypa/pip-audit/compare/v2.5.6...HEAD
-[2.5.5]: https://github.com/pypa/pip-audit/compare/v2.5.5...v2.5.6
+[2.5.6]: https://github.com/pypa/pip-audit/compare/v2.5.5...v2.5.6
 [2.5.5]: https://github.com/pypa/pip-audit/compare/v2.5.4...v2.5.5
 [2.5.4]: https://github.com/pypa/pip-audit/compare/v2.5.3...v2.5.4
 [2.5.3]: https://github.com/pypa/pip-audit/compare/v2.5.2...v2.5.3